requirements.txt
pandas
requests
mysql-connector-python
python-dotenv
python-dateutil
